% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{llenar_nas}
\alias{llenar_nas}
\title{Funcion usada para llenar los NAs en la parte de posesiones}
\usage{
llenar_nas(tabla, columna)
}
\arguments{
\item{tabla}{Tabla que recibe}

\item{columna}{Columna a llenar los NAs}
}
\value{

}
\description{
Funcion usada para llenar los NAs en la parte de posesiones
}
\examples{
llenar_nas(tabla, 'numero_jugada')
}
