% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{scrape_from_ihf}
\alias{scrape_from_ihf}
\title{Title}
\usage{
scrape_from_ihf(enlace, carpeta)
}
\arguments{
\item{enlace}{Página desde donde queremos descargar los partidos}
}
\value{

}
\description{
Title
}
\examples{
scrape_from_ihf(enlace = 'https://www.ihf.info/competitions/men/308/27th-ihf-men039s-world-championship-2021/22415/match-center/23765', carpeta = 'matches')
}
